Revert "Merge branch 'main' into develop"
This reverts commit 9f0b1e103ecee499a966019c5c118368ccc85a95, reversing
changes made to e2305b3010da3807f58c0e30fb9352964b15f127.
