+++ package.json
@@ -36,7 +36,9 @@
     "react-router-dom": "^6.20.1",
     "sockjs-client": "^1.6.1",
     "styled-components": "^6.1.1",
-    "xterm-for-react-18": "^1.0.28",
+    "styled-reset": "^4.5.1",
+    "ts-debounce": "^4.0.0",
+    "xterm": "^5.3.0",
     "yorkie-js-sdk": "^0.4.10",
     "zustand": "^4.4.7"
   },
