+++ src/index.css
@@ -1,0 +1,5 @@
+@import url('./assets/font/font.css');
+
+body {
+  font-family: 'Pretendard', sans-serif !important;
+}
