+++ src/pages/GuestEnterPage/GuestEnterPage.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 const GuestEnterPage = () => {
   return <div>GuestEnterPage</div>;
 };
