+++ src/pages/IDEPage/FileTree/FileTree.styles.ts
