+++ src/pages/IDEPage/Header/IDEHeader.style.ts
