+++ src/pages/IDEPage/Terminal/Terminal.style.ts
