+++ src/pages/IDEPage/Terminal/Terminal.tsx
