+++ src/pages/IDEPage/Terminal/TerminalTest.tsx
