--- src/pages/Login.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Login = () => {
-  return <div>Login</div>;
-};
-
-export default Login;
