+++ src/pages/MyPage/MyPage.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 const MyPage = () => {
   return <div>MyPage</div>;
 };
