--- src/pages/ProjectPage.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const ProjectPage = () => {
-  return <div>ProjectPage</div>;
-};
-
-export default ProjectPage;
