--- src/pages/SignUp.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const SignUp = () => {
-  return <div>SignUp</div>;
-};
-
-export default SignUp;
